Move scope spec.
